Various Find{,All} and import-related bits.
Find{,All}() never returns IResourceIndexEntrys with IsDeleted==true, so don't need to test for that.
Lose some unnecessary "(uint)" casts.
Various changes to dbc import for Anach.
Handle dropping/-import of a folder by doing a recursive import.

+++ S3PIDemoFE/trunk/BrowserWidget/BrowserWidget.cs
@@ -566,7 +566,7 @@ namespace S3PIDemoFE
             nameMap = new List<IResource>();
             nameMapRIEs = new List<IResourceIndexEntry>();
 
-            IList<IResourceIndexEntry> lrie = pkg.FindAll(_key => _key.ResourceType == (uint)0x0166038C);
+            IList<IResourceIndexEntry> lrie = pkg.FindAll(_key => _key.ResourceType == 0x0166038C);
             foreach (IResourceIndexEntry rie in lrie)
             {
                 try
